Change trackTimeMillis type from Int to Long in RemoteSong data class

+++ app/src/main/java/com/felpster/musicplayer/data/remote/RemoteSong.kt
@@ -84,7 +84,7 @@ data class RemoteSong(
     val trackNumber: Int,
 
     @field:Json(name = "trackTimeMillis")
-    val trackTimeMillis: Int,
+    val trackTimeMillis: Long,
 
     @field:Json(name = "country")
     val country: String,
